Added cooldown

+++ Assets/prefabs/Framework/Character.cs
@@ -22,7 +22,7 @@ public class Character : MonoBehaviour
         GetComponent<Animator>().Play(DeathStateNameHash);
     }
 
-    public virtual void HealthChanged(int newValue, int oldValue, int maxValue, GameObject Caluse)
+    public virtual void HealthChanged(float newValue, float oldValue, float maxValue, GameObject Caluse)
     {
         
     }
